finalproject-matrix/requirements.txt.txt
streamlit>=1.20.0
numpy
Pillow
opencv-python
rembg